Change unsubscribing/unregistering logic
Omnes::Bus#unsubscribe will only remove subscribers from the queue, but
it won't "unsubscribe" events. The old behavior for unsubscribing an
event is moved to Omnes::Bus#unregister. However, on top of that, it'l
unregister the event from the registry. That means it'll no longer be
possible to resubscribe to the same event (unless we re-register again).