Remove regexp subscriptions
That was an anti-pattern, plus it added a lot of complexity:

- A user might subscribe to `/order/` because they're thinking about two
`order_finalized` & `order_recalculated` events. However, a new
`stock_backordered` event could be added, and it'd break.

- The same behavior can be achieved by manually listing the events
you're interested in.

- It makes the implementation more complex. If you want to silence an
event, you can remove the associated subscriptions. However, we had to
keep matching regexp subscriptions, as they could still listen to other
events. We were using an exclusion cache to work around that.

We might consider adding a catch-all subscription type in the future.