Add feature to subscribe a single handler to all events
It was only possible to subscribe a callable object to a single event
until now. However, sometimes it's helpful to use a catch-all strategy.
For instance, in case you want to log all triggered events. Now, that's
possible with `Omnes::Bus#subscribe_to_all`:

```ruby
bus = Omnes::Bus.new
bus.subscribe_to_all do |event|
  MyLogger.call <<~MSG
    Event published at #{event.publication_time}
  MSG
end
```

To that purpose, we've abstracted the strategy by which a subscription
is called for a given event. A strategy needs to respond to `#call` and
accept the candidate event being matched. So, a nice side effect of the
new feature is the ability for a user to use custom matching strategies.
We've introduced a new `Omnes::Bus#subscribe_with_strategy` method to
leverage that:

```ruby
bus = Omnes::Bus.new
order_events_strategy = ->(candidate) { candidate.match?(/^order_/) }
bus.subscribe_with(order_events_strategy) do |event|
  # ...
end
```

The new features are also reflected on `Omnes::Subscriber` modules with
the new methods `handle_all` & `handle_with_strategy`:

```ruby
class MySubscriber
  include Omnes::Subscriber

  handle_all with: :log
  handle_with_strategy order_events_strategy, with: :order_events

  def log(event)
    # ...
  end

  def order_events(event)
    # ...
  end
end
```

We've done an important refactor on `Omnes::Subscriber` to accommodate
that. We're no longer directly adding the event name in the definition
but the strategy to be used. The definition needs to be lazily defined
to take the bus instance and check whether a given candidate is
registered. Consequently, we're losing the information of the intended
event for unknown/private methods subscription errors. It'd still be
possible to get that by iterating over all registered events. However,
that adds little value (the method name is the significant bit), and the
code is already complex enough. The `Omnes::Subscriber::Subscriptions`
result object also needed to be adapted to the indirect subscription to
specific event names.