Allow custom callback builder for subscribers
This is very useful to build async adapters. E.g.:

```ruby
class MySubscriber
  include Omnes::Subscriber
  include Sidekiq::Job

  handle :foo, with: lambda do |instance|
    lambda do |event|
      instance.class.perform_async(event.object.id)
    end
  end

  def perform(object_id)
    # ...
  end
```

We make the builder that only takes a method on the instance the default
when a `Symbol` is given instead of a lambda.

Being the generated subscriptions no longer hard-attached to methods,
there's no reason to return the special collection of method-backed
subscriptions. This is a loose in introspection that we might fix later,
although we can also argue whether it makes sense to build abstractions
for specialized collections of subscriptions.

We're also removing the duplicate subscriptions check, as some
strategies could allow multiple subscriptions.