Allow publishing any instance responding to `#name`
Until now, an event publication provided a name and an associated
payload meant to be consumed by its subscriptions. That was converted
under the hood into an `Omnes::Event` instance. From now on, this
behavior is still supported. However, an `Omnes::UnstructuredEvent`
instance is returned stead.

The now more generic `Omnes::Event` is reserved to allow publishing any
instance of a subclass inheriting it. When you publish an instance,
there's no longer the need to provide the event name or the payload.
Subscribers will have access to that instance and everything it
provides. The benefit of that is having a structured and centralized
definition of the event, where developers can inspect instead of looking
for runtime generated payloads. It also opens gets to implementing
storage stragies for the published events.

The publication of instances is not limited to `Omnes::Event`'s
instances. It can, in fact, be anything responding to `#name`. On an
`Omnes::Event` object, the name equals the underscored class name,
replacing module delimiters (i.e., `::`) with underscores.

Events can now have different shapes, so the publication time and caller
location are no longer added to id. Instead, that information is
collected in the `Omnes::Publication` instance.

Example:

```ruby
class MyEvent < Omnes::Event
  attr_reader :id

  def initialize(id)
    @id = id
  end
end

bus = Omnes::Bus.new
bus.register(:my_event)
bus.subscribe(:my_event) do |event|
  puts event.id
end
bus.publish(MyEvent.new(1))
```