Introduce Sidekiq adapter
To begin with, we're renaming the `CallbackBuilder` module to `Adapter`,
as the new one is more user-friendly to use. We add then a new adapter
to be used to run subscription callbacks as Sidekiq's background jobs.

We provide the result value of a `payload` method in the event to the
required `#perform` method in the job. That plays OOTB with our default
`UnstructuredEvent`. Users need to make the payload serializable
according to Sidekiq's needs.

```ruby
class MySubscriber
  include Omnes::Subscriber
  include Sidekiq::Job

  handle :my_event, with: Adapter::Sidekiq

  def perform(payload)
    puts payload # "{ "foo"=>"bar" }"
  end
end
bus = Omnes::Bus.new
bus.register(:my_event)
bus.publish(:my_event, "foo" => "bar")
```

Analogous to Sidekiq's `.perform_in` method, we give the option to delay
the job by the given interval.

```ruby
handle :my_event, with: Adapter::Sidekiq.in(60)
```