Allow uniquely identifying subscriptions
We add an `id:` param to all the methods to create subscriptions. We can
use it to fetch a given subscription from the bus when a reference to it
is no longer available. Example:

```ruby
subscription = bus.subscribe(:foo, id: :foo_subscription) { do_something }
bus.subscription(:foo) == subscription # => true
```

That's very useful as it allows us to make the event bus the only global
piece in the code and use it for testing purposes like in the
`Omnes::Bus#performing_only` method.