Support Ruby 2.5 & 2.6
Omnes was born as an extraction from
[Solidus](https://github.com/solidusio/solidus) and it's used there to
provide pub/sub support. Current Solidus still supports Ruby 2.5 & 2.6.
We'll deprecate here as soon as they're deprecated on Solidus (which
should happen in next major).

An edge case that won't work on those versions is passing a serialized
payload when publishing an unstructured event, as it'll require the
kwargs keys to be Symbols. That's something not used on Solidus, so we
don't make any changes to support it. For those cases, the workaround is
publishing an event instance with the `#payload` method.